+++ src/weeimport/weathercatimport.py
@@ -284,6 +284,12 @@ class WeatherCatSource(weeimport.Source):
         print(_msg)
         log.info(_msg)
         self.print_map()
+        if self.update:
+            _msg = "Imported records will overwrite existing database records."
+        else:
+            _msg = "Imported records will not overwrite existing database records."
+        print(_msg)
+        log.info(_msg)
         if self.calc_missing:
             print("Missing derived observations will be calculated.")
         if not self.UV_sensor: